Update to 6-size

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long int __i;
 	float f;
 
-	printf("Size of a char: %d byte(s)\n", (unsigned long)sizeof(char c));
-	printf("Size of an int: %d byte(s)\n", (unsigned long)sizeof(int i));
-	printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(long int _i));
-	printf("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(long long int __i));
-	printf("Size of a float: %d byte(s)\n", (unsigned long)sizeof(float f));
+	printf("Size of a char: %d byte(s)\n", sizeof(char c));
+	printf("Size of an int: %d byte(s)\n", sizeof(int i));
+	printf("Size of a long int: %d byte(s)\n", sizeof(long int _i));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(long long int __i));
+	printf("Size of a float: %d byte(s)\n", sizeof(float f));
 	return (0);
 }
